requirements.txt
pandas
numpy
scikit-learn
pymongo
python-dotenv
streamlit
#-e .